custom layout manager

+++ shop/data/src/main/java/com/example/data/model/ResPlant.kt
@@ -10,7 +10,7 @@ data class ResPlant(
     val category: String,
     val imageUrl: String,
     val videoUrl: String,
-    val price: Int,
-    val approxHeight: Int,
+    val price: Long,
+    val approxHeight: Long,
     val maintenance: String
 )
